requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
deep-translator==1.11.4
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.6
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
